Switch to Value and AssocMap newtypes; Remove old types
