+++ plutarch-ledger-api/plutarch-ledger-api.cabal
@@ -47,7 +47,6 @@ library
   import:          common-lang
   exposed-modules:
     Plutarch.LedgerApi.AssocMap
-    Plutarch.LedgerApi.AssocMapNew
     Plutarch.LedgerApi.Interval
     Plutarch.LedgerApi.Utils
     Plutarch.LedgerApi.V1
@@ -70,7 +69,9 @@ library
     Plutarch.LedgerApi.V3.Contexts
     Plutarch.LedgerApi.V3.MintValue
     Plutarch.LedgerApi.V3.Tx
+    Plutarch.LedgerApi.Value.AssetClass
     Plutarch.LedgerApi.Value.CurrencySymbol
+    Plutarch.LedgerApi.Value.Lovelace
     Plutarch.LedgerApi.Value.TokenName
 
   build-depends:
